LayoutTests/imported/w3c/web-platform-tests/clipboard-apis/async-promise-write-blobs-read-blobs.https-expected.txt

Body needed for test_driver.click()

FAIL Verify write and read clipboard (multiple types) with promise to Blobs assert_equals: expected 2 but got 1
